Add my solution to the coding exercise
This code is unchanged since I originally wrote it, back in September
of 2006. It appears to still render correctly in Safari for Mac, but
the JavaScript email validation doesn't appear to work anymore. I'm
leaving it broken for archival reasons.

It's useful to have a copy of my coding exercise to compare to. It's
almost certainly not perfect, and nothing like how I'd do it today, but
it might be interesting to someone.